Issue #2 Removed the small classes that rendered a small amount of code. Trying to move Users to the additional information.
