--- src/components/Additional/AdditionalChildren/Locations.jsx
@@ -1,20 +1,0 @@
-/**
- * This is how each location will be displayed.
- */
-
-import React, { Component } from "react";
-
-class Location extends Component {
-  render() {
-    const locationStyle = {
-      marginTop: "5px"
-    };
-    return (
-      <li className="list-group-item" style={locationStyle}>
-        {this.props.location}
-      </li>
-    );
-  }
-}
-
-export default Location;
